+++ winsource/winsource.def
@@ -1,0 +1,12 @@
+;===========================================================================
+;  Copyright (c) 1992-2002  Microsoft Corporation.  All Rights Reserved.
+;===========================================================================
+
+LIBRARY     Ball.dll
+
+EXPORTS
+            DllMain                 PRIVATE
+            DllGetClassObject       PRIVATE
+            DllCanUnloadNow         PRIVATE
+            DllRegisterServer       PRIVATE
+            DllUnregisterServer     PRIVATE
